requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pyjwt==2.8.0
python-dotenv==1.0.1
werkzeug==3.0.1
beautifulsoup4==4.12.3
requests==2.31.0 